dividends/requirements.txt
pyspark==3.5.0
pytz==2025.2
Requests==2.32.5
google-cloud-bigquery==3.40.1
google-cloud-storage==3.9.0
google-cloud-logging==3.13.0
google-cloud-secret-manager==2.26.0
